Account for cmd window in save, use API for buf contents
When the cmd window is active (e.g. spawned by q: or q/), there are
several locks that restrict actions (:h E11):

The command-line window is not a normal window. It is not possible to move to
another window or edit another buffer. All commands that would do this are
disabled in the command-line window.

In particular, setting the current tabpage/win/buf does not work
(setting the tabpage is part of saving regular sessions) and calling
:wundo yields the same error. I'm unsure if this is because of the
nvim_buf_call or wundo itself, but the error is traced back to the wundo
line specifically.
Since we only switch the active tabpage to save the otherwise
inaccesible tab-scoped `cmdheight` option, just skip switching it and
warn about the issue.

Relying on

> vim.cmd.w({save_file, bang = true,
>   mods = { keepalt = true, noautocmd = true, silent = true },
> })

for writing the buffer contents seemed to cause side effects, hence we
now read the buffer contents via the API and dump them ourselves, which
avoids these issues (unsure about tradeoffs e.g. related to encoding,
time will tell).