fix: Restoration after buffer in active window was filtered
When the active window contains a buffer that was filtered, no current
window can be set after restoration. That's not an issue if the last
restored tabpage has at least one other window to display, otherwise the
user is confronted with an empty, unlisted buffer after loading the
session.

The previous logic resulted in unexpected switching of buffers when
the tabpage had more windows. I think the proper fix would be to also
save the active tabpage and at least switch to that one.