fix: Jumplist breakage in active window with per-session history
Executing `wshada` not only seems to set the `"` mark in all visible
windows, but also breaks the jumplist in the active window if the
current position is not past the last entry.

Workaround by opening a temporary active window.

nvim 0.12 will likely not exhibit this behavior because it respects
`'0` as documented and thus skips the call to `shada_init_jumps`.
Untested at the moment though.