fix: Improve restoration when active buffer is filtered
When an active buffer is filtered, its window is removed from the layout
as well. If that window is the only window in the current tabpage, the
active tabpage would be lost, defaulting to the last one instead.
If the last tabpage was affected by this, the user would be presented
with an empty scratch buffer after restoration.

The reworked logic was inherited from Resession and intends to solve the
latter situation, but was broken when switching to indicating an empty
window layout via an empty table instead of `false` and deferring the
restoration of invisible buffers, causing (if no other buffer was
visible at the time of saving the snapshot) no known `last_bufnr` and
(if the affected tabpage was not the last one) a partially restored
buffer to be shown (since all buffers in inactive tabpage windows are
restored later as well).

This commit ensures:
* If active window is filtered, always ensure that the active tabpage is
  kept regardless.
* If the active tabpage is affected, some buffer is shown in completely
  initialized state.
* We still don't crash when loading a Resession snapshot.

Improvements:
It would be more expected to instead load the last non-filtered buffer
that was actually shown in the window (either via alternative buffer or
jumplist).
Consider keeping all windows intact, even if their buffers are filtered.
We save a lot more data than Resession does, so we e.g. lose jumplists
of affected windows if they happen to show the wrong buffer during save.
If not, consider handling the buffer selection during save, where we
still have this info.