feat: Restore relevant hidden bufs earlier
Previously, all hidden buffers were restored in a delayed step, which
could cause minor issues when switched to immediately (e.g. alternate
buffers). This included non-current tabpage windows.

This commit forces restoration for all buffers that are displayed in any
window and also all window-specific alternate buffers. Other hidden
buffers are still restored later, either when a `CursorHold[I]` event
happens or after some timeout. This timeout has been halved to 500ms for
the first hidden buffer, then increases by 30ms for each following one.

I considered dropping the autocmd completely in favor of the deferred
functions acting as a timeout, but decided against it because usually
the restoration is quick and should not be noticeable. If this ever
becomes an issue, it can still be dropped.

If an "unrestored" buffer is loaded into a window before its restoration
has been requested, it usually does not cause much of an issue, but
plugins and syntax highlighting can be broken and the last cursor
position is not restored. If the buffer is left and re-entered after, it
is reinitialized as usual (can cause the cursor to jump).