feat: Fully restore all visible buffers/all windows immediately
... instead of waiting for the window to be focused. This avoids missing
annotations from plugins, for example.

Might also fix an issue where two windows show the same buffer and one
of them is focused and thus re-edited, resetting the cursor in the
second window (unsure if this is only triggered by the rapid re-editing
or if this was an issue before though).