feat: Add basic neogit plugin
This plugin supports restoration of status and single commit views.
I explored restoring log views as well, but as noted in the comments at
the bottom this proved very hard to do and thus brittle.

The only view that could be restored somewhat reliably was the one
opened by the inbuilt `NeogitLogCurrent` command.
Discerning this view from the others would be possible by running a
simple pcall(popup, "get_arguments") and checking if it does not error
(other popups need method-style), but it's still not targetable to a
window and relies on debug.getupvalue.